production-center/core-application/TikTok/requirements.txt
requests
pandas
loguru
google-cloud-bigquery
google-cloud-secret-manager
google-auth
functions-framework
PyYAML
